backend/requirements.txt
librosa>=0.10.0
numpy>=1.26.0
scipy>=1.11.0
soundfile>=0.12.1
sounddevice>=0.4.6
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
websockets>=12.0

# Video search dependencies
openai-whisper>=20231117
sentence-transformers>=2.2.2
chromadb>=0.4.18
moviepy>=1.0.3
torch>=2.0.0
torchaudio>=2.0.0